feat(Cypress): refactor after review
